requirements.txt
selenium==3.141.0
beautifulsoup4==4.9.3
Pillow==8.3.1
